Scorecard: Refactor score card components
