+++ src/responsibleai/rai_analyse/_score_card/box_plot_explain.png
