+++ src/responsibleai/rai_analyse/create_score_card.py
@@ -7,14 +7,18 @@ import re
 from azureml.core import Run
 from responsibleai import __version__ as responsibleai_version
 
-from _rai_insight_data import RaiInsightData, PdfDataGen
+from _score_card._rai_insight_data import RaiInsightData, PdfDataGen
 
-from generate_pdf import (
-    RegressionComponents,
-    ClassificationComponents,
-    to_pdf,
-    get_full_html,
-)
+# from _score_card.generate_pdf import (
+#     RegressionComponents,
+#     ClassificationComponents,
+#     to_pdf,
+#     get_full_html,
+# )
+
+from _score_card.common_components import to_pdf, get_full_html
+import _score_card.regression_components as RegressionComponents
+import _score_card.classification_components as ClassificationComponents
 
 from constants import DashboardInfo, PropertyKeyValues, RAIToolType
 from rai_component_utilities import load_dashboard_info_file
